+++ verify/model_otp_device_configuration.go
@@ -40,6 +40,8 @@ func NewOTPDeviceConfiguration(verify EnumVerify) *OTPDeviceConfiguration {
 // but it doesn't guarantee that properties required by API are set
 func NewOTPDeviceConfigurationWithDefaults() *OTPDeviceConfiguration {
 	this := OTPDeviceConfiguration{}
+	var verify EnumVerify = ENUMVERIFY_DISABLED
+	this.Verify = verify
 	return &this
 }